changed models.py to output humour probability
-deleted HumourRegconitionModel
- finalized  that our model class to be stored at team7/aiap-team-7-project-jokebot/src/aiap_team_7_project_jokebot/modeling/models.py
